feat: refactor notification invalidation logic to include pathname handling and toast notifications for direct messages
